+++ src/main/java/warmup/two/WarmupTwoSecondSet.java
@@ -38,4 +38,27 @@ public class WarmupTwoSecondSet {
 
         return count;
     }
+
+    /**
+     * Given an array of ints, return true if one of the first 4 elements in the array is a 9. The
+     * array length may be less than 4.
+     *
+     * @param nums
+     * @return
+     */
+    public boolean arrayFront9(int[] nums) {
+        int numsLen = nums.length;
+
+        if (numsLen > 4) {
+            numsLen = 4;
+        }
+
+        for (int i = 0; i < numsLen; i++) {
+            if (nums[i] == 9) {
+                return true;
+            }
+        }
+
+        return false;
+    }
 }